Major refactoring.
* The old strategy consisted of each child component remembering its
children components. However, it was difficult to work with it. The
strategy is now to have the Canvas store the trees containing the
properties and tags, and each children having its own as well. The
draggables don't do much now besides rendering themselves and handling a
few mouse events. All the functionality was moved to the Canvas.
* Some changes to UI/Element.hs. It now has a function toHTML which
transforms HTML functions to ComponentHTML.
* Some bugs still persist, for example, the trees sometimes erasing the
wrong thing or the points not being relative to their parents.