Added Array type (kinda WIP) and fixed JSON naming
Changes to FromJSON instances:
* They all follow a pattern like ModuleName.TypeName, for example,
Foundation.EditorPOST.
* FromJSON and ToJSON instances are now in the deriving clause of the
data declarations.

Changes to low-code.cabal:
* Added DeriveAnyClass to library.
* Removed Language.LowCode.Logic.{Prim,Types} and merged them in
Analyzer.

Changes to Common:
* Added a Double2 = (Double, Double) type.

Changes to JS AST:
* Added an Array type.

Changes to Logic AST:
* Added an Array type.
* Renamed Metadata to Environment.
* Added a Metadata type, currently only containing a Double2
representing the position. This will later be added to the Analyzer to
display better error messages.
* Merged Variable and VariableType from the Types module (now deleted)
to the AST module.
* Fixed a small sanity check for operators, where it would say an unary
operator is unknown, even if the operator is binary.
* Added Array parser. Keep in mind that it still doesn't work very well
with empty arrays, as it deduces the inner type as Unit.

Changes to Logic Analyzer:
* The Prim module was merged to the Analyzer.
* Added an evalAnalyzerT function, analog to the execAnalyzerT function
already present.
* Now analyzes array types. Keep in mind that it still doesn't work well
for empty arrays (see last note in *Changes to Logic AST*).