UI integration first test.
Changes to low-code.cabal:
* Add lib-Werror flag, so warnings can optionally be turned into errors.
This makes testing easier.
* Added a common shared-properties stanza containing some common
properties used in other stanzas.
* Added ApplicativeDo extension.
* Added containers, data-default-class, HUnit and QuickCheck
dependencies to the test-suite test. Although The two latter libraries
are not used yet, I plan to make proper use of them in the future.
* Renamed main-is from non-existing Tests.hs to the newly added Main.hs.

Changes to Bundles:
* Added extraJsFiles to BundleCssHtmlLogic so it's impossible to include
additional JS files in the project.

Changes to HTML AST:
* When including scripts, it will now include a "text/javascript" type.

Changes to Logic AST and JS AST:
* Changed Assign to use an Expression rather than a Name.

Changes to JS AST:
* Removed symbol table, as we assume that the symbols were already
validated by the Logic Analyzer.
* Generates Assign according to the change in the AST.

Changes to Logic Analyzer:
* Analyzes Assign according to the change in the AST.

Changes to Logic to JavaScript converter:
* Generates Assign according to the changes in the ASTs.

Changes to test:
* Renamed AST.hs to Main.hs.
* Added a simple test containing a counter with buttons.