Added better JSON instances to Logic Expression.
Changes to Bundles:
* Replaced usage of lToJs to convert.

Changes to Logic:
* Export JavaScriptConverter module.

Changes to Logic AST:
* Fixed FromJSON AST error message so it's consistent with other
messages.
* FromJSON Expression now checks whether the received Value is a String
or Object. If it's a string, it just runs the parser. Otherwise, it
parses an actual JSON Object. If none of those two is input, then it
displays an error message.
* ToJSON Expression now generates an appropriate JSON Object. It never
generates a "human-writable" format.
* Update ToJSON Variable to use LambdaCase rather than pattern matching
on function, for consistency.