Changed JS Function to be an Expression rathen than AST.
Changes to JS AST:
* Moved Function from AST to Expression.
* Reorganized Expression alphabetically.

Changes to JS Codegen:
* Moved Function from javaScriptCodegenInternal to genExpression.
* Reorganized genExpression alphabetically.

Changes to Logic:
* Disabled dodgy-exports and unused-imports warnings. There seems to be
a bug in which somethung that uses convert makes GHC complain about the
function not existing when the module is not imported. However, GHC
thinks it's an empty export, and so explicitly disable these warnings.

Changes to Logic to JS converter:
* Removed lToJs.
* Logic.Start now generates JS.Function inside a JS.Expression.