Added unreachable statement analysis.
Changes to JS Codegen:
* Fixed an issue where records would emit commas instead of semicolons
between fields in records.

Changes to Logic Analyzer:
* Variables that begin with _ are no longer considered unused.
* Variables shadowed in name patterns now throw errors.
* Add analysis to find unreachable statements, and throw an error in
case a function has no return and its return type is not Unit.
* Fixed an issue where it would not be possible to index TextTypes.
* When typechecking, expressions will now use the expected type instead
of the inferred type.

Changes to Logic Module:
* Module name expects a semicolon now.

Changes to Logic JSON module:
* Move unreachable statements to inside match patterns, effectively
using the fix made in Logic Analyzer (item 3). Previously, it was
defined like this as a workaround.